upated dataset

+++ python/dgl/data/reddit.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import
 
 import scipy.sparse as sp
 import numpy as np
+import networkx as nx
 import dgl
 import os, sys
 from .utils import download, extract_archive, get_download_dir, _get_dgl_url
@@ -19,6 +20,7 @@ class RedditDataset(object):
         extract_archive(zip_file_path, extract_dir)
         # graph
         coo_adj = sp.load_npz(os.path.join(extract_dir, "reddit{}_graph.npz".format(self_loop_str)))
+        #self.graph = nx.from_scipy_sparse_matrix(coo_adj, create_using=nx.DiGraph())
         self.graph = dgl.DGLGraph(coo_adj, readonly=True)
         # features and labels
         reddit_data = np.load(os.path.join(extract_dir, "reddit_data.npz"))
@@ -28,9 +30,9 @@ class RedditDataset(object):
         # tarin/val/test indices
         node_ids = reddit_data["node_ids"]
         node_types = reddit_data["node_types"]
-        self.train_mask = (node_types == 1)
-        self.val_mask = (node_types == 2)
-        self.test_mask = (node_types == 3)
+        self.train_mask = (node_types == 1).astype(int)
+        self.val_mask = (node_types == 2).astype(int)
+        self.test_mask = (node_types == 3).astype(int)
 
         print('Finished data loading.')
         print('  NumNodes: {}'.format(self.graph.number_of_nodes()))
